+++ test/ChickenCheck.WebTests/Program.fs
@@ -8,22 +8,23 @@ open canopy.runner.classic
 open canopy.types
    
 canopy.configuration.failFast := true
+canopy.configuration.firefoxDir <- "/Applications/Firefox\ Developer\ Edition.app"
 
 [<EntryPoint>]
 let main args =
 
-    let rootUrl = args |> Array.tryHead |> Option.defaultValue "http://localhost:8085"
+    let rootUrl = args |> Array.tryHead |> Option.defaultValue "https://localhost:8085"
     let failIfAnyWipTests = if args.Length = 2 then Convert.ToBoolean(args.[1]) else false
     canopy.configuration.failIfAnyWipTests <- failIfAnyWipTests
     start BrowserStartMode.ChromeHeadless
-    // start BrowserStartMode.Chrome
+//    start BrowserStartMode.Chrome
     resize (1920, 1080)
     
     "Root url redirects to chickens page" &&& fun _ ->
         let today = DateTime.Today
-        let chickensPage = ChickensPage.url rootUrl today
+        let eggsPage = EggsPage.url rootUrl today
         url rootUrl
-        onn chickensPage
+        onn eggsPage
     
     Chickens.all rootUrl